ioctl: export nvme_{g,s}et_features_iocs_profile()
nvme_get_features_iocs_profile() and nvme_set_features_iocs_profile()
are defined but not exorted in libnvme.so.
nvme_set_features_iocs_profile()'s prototype was also removed,
so add it back.
IOCSI is a 9-bit value, so fix its bitmask and change its type to u16.

+++ src/nvme/ioctl.c
@@ -887,7 +887,7 @@ int nvme_set_features_write_protect(int fd, enum nvme_feat_nswpcfg_state state,
 				   false, result);
 }
 
-int nvme_set_features_iocs_profile(int fd, __u8 iocsi, bool save)
+int nvme_set_features_iocs_profile(int fd, __u16 iocsi, bool save)
 {
 	__u32 value = NVME_SET(iocsi, FEAT_IOCSP_IOCSCI);
